Move input logic to input.rs
